cleaned up code
much cleaner code

+++ src/main/java/service/FriendRequestService.java
@@ -1,53 +1,46 @@
-package models;
+package service;
 
-import domain.FriendRequest;
+import models.FriendRequest;
 import enums.ChangeEvent;
 import exceptions.ValidationException;
-import javafx.collections.FXCollections;
-import javafx.collections.ObservableList;
 import repo.DbFriendRequestRepo;
 import utils.Tuple;
-import utils.observer.NotificationHandler;
 import utils.observer.Observable;
+import java.util.ArrayList;
+import java.util.List;
 import java.util.Optional;
 
 
-public class FriendRequestModel extends Observable {
+public class FriendRequestService extends Observable {
 
     private final DbFriendRequestRepo friendRequestRepo;
     private FriendRequest lastFriendRequestSaved;
 
-    public FriendRequestModel(DbFriendRequestRepo friendRequestRepo) {
+    public FriendRequestService(DbFriendRequestRepo friendRequestRepo) {
 
         this.friendRequestRepo = friendRequestRepo;
-        this.addObserver(NotificationHandler.getInstance());
     }
 
     public Optional<FriendRequest> findOne(Long id1, Long id2){
+        if(id1 == null || id2 == null || id1<0 || id2<0){
+            throw new ValidationException("Invalid ids");
+        }
         return friendRequestRepo.findOne(new Tuple<>(id1,id2));
     }
 
-//    public ObservableList<FriendRequest> findAll(){
-//        ObservableList<FriendRequest> list = FXCollections.observableArrayList();
-//        for(FriendRequest fr:friendRequestRepo.findAll()){
-//            list.add(fr);
-//        }
-//        return list;
-//    }
 
-    public ObservableList<FriendRequest> findFriendRequestsOf(Long id){
-        ObservableList<FriendRequest> list = FXCollections.observableArrayList();
-        for(FriendRequest fr:friendRequestRepo.findFriendRequestsOf(id)){
-            list.add(fr);
+    public List<FriendRequest> findFriendRequestsOf(Long id){
+        if(id == null || id<0){
+            throw new ValidationException("Invalid ids");
         }
+        List<FriendRequest> list = new ArrayList<>();
+        friendRequestRepo.findFriendRequestsOf(id).forEach(list::add);
         return list;
     }
 
-    public ObservableList<FriendRequest> findSentFriendRequests(Long id){
-        ObservableList<FriendRequest> list = FXCollections.observableArrayList();
-        for(FriendRequest fr:friendRequestRepo.findSentFriendRequests(id)){
-            list.add(fr);
-        }
+    public List<FriendRequest> findSentFriendRequests(Long id){
+        List<FriendRequest> list = new ArrayList<>();
+        friendRequestRepo.findSentFriendRequests(id).forEach(list::add);
         return list;
     }
 
@@ -57,8 +50,8 @@ public class FriendRequestModel extends Observable {
         notifyObservers(ChangeEvent.SENT_FRIEND_REQUEST);
     }
     public void delete(Long id1, Long id2){
-        if(id1<0 || id2<0){
-            throw new ValidationException("Id's cannot be negative");
+        if(id1 == null || id1<0 || id2 == null || id2<0){
+            throw new ValidationException("Invalid ids!");
         }
         friendRequestRepo.delete(new Tuple<>(id1,id2));
         notifyObservers(ChangeEvent.FRIEND_REQUEST_DATA);
